Quick filtering (#1)
This changes the UX for applying filters. Instead of displaying a bar of filters above the DashboardTable, every column has a popover for applying an ad-hoc filter. It also adds the ability to apply a column filter based on a list of values. These two changes make it much easier to explore data.