Shareable dashboard URLs (#7)
This encodes the current order setting and filters into a URL parameter, so links to a dashboard will automatically apply these settings on load. This makes Datasite links more shareable, because a link can include a "data story" to illustrate something specific about a dataset.

To serialize the state, we create a small payload in an object, then encode to the msgpack binary format. This format includes non-ASCII characters, so we base64-encode the result to use URL-safe characters. This way, we make the state as compact as possible, so links are shorter.